Move server setup to handler, add altair path
