+++ src/pages/api/altair/[[...any]].ts
@@ -1,0 +1,3 @@
+import { graphqlHandler } from "../../../graphql/handler";
+
+export default graphqlHandler;
